+++ tests/pkg/fault-trigger/manager/vm.go
@@ -14,12 +14,7 @@
 package manager
 
 import (
-	"fmt"
-	"os/exec"
 	"strings"
-	"sync"
-
-	"github.com/golang/glog"
 )
 
 type VMManager interface {
@@ -27,90 +22,6 @@ type VMManager interface {
 	ListVMs() ([]*VM, error)
 	StopVM(*VM) error
 	StartVM(*VM) error
-}
-
-type VirshVMManager struct {
-	*sync.RWMutex
-	vmCache map[string]string
-}
-
-func (m *VirshVMManager) Name() string {
-	return "virsh"
-}
-
-// ListVMs lists vms
-func (m *VirshVMManager) ListVMs() ([]*VM, error) {
-	shell := fmt.Sprintf("virsh list --all")
-	cmd := exec.Command("/bin/sh", "-c", shell)
-	output, err := cmd.CombinedOutput()
-	if err != nil {
-		glog.Errorf("exec: [%s] failed, output: %s, error: %v", shell, string(output), err)
-		return nil, err
-	}
-	vms := m.parserVMs(string(output))
-	return vms, nil
-}
-
-// StopVM stops vm
-func (m *VirshVMManager) StopVM(v *VM) error {
-	shell := fmt.Sprintf("virsh destroy %s", v.Name)
-	cmd := exec.Command("/bin/sh", "-c", shell)
-	output, err := cmd.CombinedOutput()
-	if err != nil {
-		glog.Errorf("exec: [%s] failed, output: %s, error: %v", shell, string(output), err)
-		return err
-	}
-
-	glog.Infof("virtual machine %s is stopped", v.Name)
-
-	return nil
-}
-
-// StartVM starts vm
-func (m *VirshVMManager) StartVM(v *VM) error {
-	shell := fmt.Sprintf("virsh start %s", v.Name)
-	cmd := exec.Command("/bin/sh", "-c", shell)
-	output, err := cmd.CombinedOutput()
-	if err != nil {
-		glog.Errorf("exec: [%s] failed, output: %s, error: %v", shell, string(output), err)
-		return err
-	}
-
-	glog.Infof("virtual machine %s is started", v.Name)
-
-	return nil
-}
-
-// example input:
-//  Id    Name                           State
-// ----------------------------------------------------
-// 6     vm2                            running
-// 11    vm3                            running
-// 12    vm1                            running
-// -     vm-template                    shut off
-func (m *VirshVMManager) parserVMs(data string) []*VM {
-	data = stripEmpty(data)
-	lines := strings.Split(data, "\n")
-	var vms []*VM
-	for _, line := range lines {
-		fields := strings.Split(line, " ")
-		if len(fields) < 3 {
-			continue
-		}
-		if !strings.HasPrefix(fields[1], "vm") {
-			continue
-		}
-
-		if strings.HasPrefix(fields[1], "vm-template") {
-			continue
-		}
-		vm := &VM{
-			Name:   fields[1],
-			Status: fields[2],
-		}
-		vms = append(vms, vm)
-	}
-	return vms
 }
 
 func stripEmpty(data string) string {
@@ -128,82 +39,4 @@ func stripEmpty(data string) string {
 		stripLines = append(stripLines, stripLine)
 	}
 	return strings.Join(stripLines, "\n")
-}
-
-type QMVMManager struct {
-	*sync.RWMutex
-	subNet  string
-	vmCache map[string]string
-}
-
-func (qm *QMVMManager) Name() string {
-	return "qm"
-}
-
-func (qm *QMVMManager) ListVMs() ([]*VM, error) {
-	shell := fmt.Sprintf("qm list")
-	cmd := exec.Command("/bin/sh", "-c", shell)
-	output, err := cmd.CombinedOutput()
-	if err != nil {
-		glog.Errorf("exec: [%s] failed, output: %s, error: %v", shell, string(output), err)
-		return nil, err
-	}
-	vms := qm.parserVMs(string(output))
-	return vms, nil
-}
-
-func (qm *QMVMManager) StartVM(vm *VM) error {
-	shell := fmt.Sprintf("qm start %s", vm.Name)
-	cmd := exec.Command("/bin/sh", "-c", shell)
-	output, err := cmd.CombinedOutput()
-	if err != nil {
-		glog.Errorf("exec: [%s] failed, output: %s, error: %v", shell, string(output), err)
-		return err
-	}
-
-	glog.Infof("virtual machine %s is started", vm.Name)
-
-	return nil
-}
-
-func (qm *QMVMManager) StopVM(vm *VM) error {
-	shell := fmt.Sprintf("qm stop %s", vm.Name)
-	cmd := exec.Command("/bin/sh", "-c", shell)
-	output, err := cmd.CombinedOutput()
-	if err != nil {
-		glog.Errorf("exec: [%s] failed, output: %s, error: %v", shell, string(output), err)
-		return err
-	}
-
-	glog.Infof("virtual machine %s is stopped", vm.Name)
-	return nil
-}
-
-// example input:
-// VMID NAME                 STATUS     MEM(MB)    BOOTDISK(GB) PID
-// 101 CentOS7600           stopped    1024              32.00 0
-// 104 to20190915-tongmu    running    8192             500.00 34863``
-func (qm *QMVMManager) parserVMs(data string) []*VM {
-	vms := []*VM{}
-	data = stripEmpty(data)
-	lines := strings.Split(data, "\n")
-	var startIndex int
-	for i, line := range lines {
-		if strings.Contains(line, "VMID") {
-			startIndex = i + 1
-			break
-		}
-	}
-	vmLines := lines[startIndex:]
-	for _, line := range vmLines {
-		fields := strings.Split(line, " ")
-		if len(fields) >= 6 {
-			vm := &VM{
-				Name:   fields[0],
-				Status: fields[2],
-			}
-			vms = append(vms, vm)
-		}
-	}
-	return vms
 }
